refactor: page 컴포넌트 방식 변경
