--- apps/web/src/pages/fit-analysis/fit-analysis-page.css.ts
@@ -1,8 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const container = style({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '2.4rem',
-  paddingBottom: '5.4rem',
-});
