+++ apps/web/src/pages/fit-analysis/fit-analysis-page.tsx
@@ -1,18 +1,11 @@
-import { JobRecommendationList, VisaEligibility } from '@widgets/fit-analysis';
-import { Section } from '@shared/ui';
-
-import { container } from './fit-analysis-page.css';
+import { UploadSection, VisaEligibilitySection } from '@widgets/fit-analysis';
 
 const FitAnalysisPage = () => {
   return (
-    <div className={container}>
-      <Section title="Check visa eligibility">
-        <VisaEligibility />
-      </Section>
-      <Section title="Find Your Perfect Job Match">
-        <JobRecommendationList />
-      </Section>
-    </div>
+    <>
+      <VisaEligibilitySection />
+      <UploadSection />
+    </>
   );
 };
