+++ apps/web/src/widgets/fit-analysis/index.ts
@@ -4,6 +4,8 @@ export { VISA_CHECKLIST_D10 } from './constants/visa-checklist';
 export { VISA_CHECKLIST_E7 } from './constants/visa-checklist';
 export { default as JobRecommendationList } from './ui/job-recommanded-list/job-recommended-list';
 export { default as UploadBox } from './ui/upload-box/upload-box';
+export { default as UploadSection } from './ui/upload-section/upload-section';
 export { default as VisaChecklist } from './ui/visa-checklist/visa-checklist';
 export { default as VisaEligibility } from './ui/visa-eligibility/visa-eligibility';
+export { default as VisaEligibilitySection } from './ui/visa-eligibility-section/visa-eligibility-section';
 export { default as VisaTabList } from './ui/visa-tab-list/visa-tab-list';
