+++ apps/web/src/widgets/fit-analysis/ui/upload-section/upload-section.css.ts
@@ -1,0 +1,5 @@
+import { style } from '@vanilla-extract/css';
+
+export const container = style({
+  padding: '2.4rem 0 4rem 0',
+});
