+++ apps/web/src/widgets/fit-analysis/ui/visa-eligibility-section/visa-eligibility-section.tsx
@@ -1,0 +1,12 @@
+import { VisaEligibility } from '@widgets/fit-analysis';
+import { Section } from '@shared/ui';
+
+const VisaEligibilitySection = () => {
+  return (
+    <Section title="Check visa eligibility">
+      <VisaEligibility />
+    </Section>
+  );
+};
+
+export default VisaEligibilitySection;
